+++ src/org/lukasj/idea/torquescript/annotator/TSMethodCallAnnotator.kt
@@ -1,20 +1,15 @@
 package org.lukasj.idea.torquescript.annotator
 
 import com.intellij.lang.annotation.AnnotationHolder
-import com.intellij.lang.annotation.HighlightSeverity
-import com.intellij.openapi.editor.markup.TextAttributes
 import com.intellij.psi.PsiElement
-import com.intellij.psi.util.elementType
 import org.lukasj.idea.torquescript.editor.TSSyntaxHighlightingColors
 import org.lukasj.idea.torquescript.engine.EngineApiService
-import org.lukasj.idea.torquescript.psi.TSTypes
 import org.lukasj.idea.torquescript.psi.impl.TSAccessorChainImpl
 import org.lukasj.idea.torquescript.psi.impl.TSFunctionCallExpressionElementImpl
 import org.lukasj.idea.torquescript.reference.ReferenceUtil
 import org.lukasj.idea.torquescript.psi.impl.TSFunctionStatementElementImpl
 import org.lukasj.idea.torquescript.psi.impl.TSFunctionType
 import org.lukasj.idea.torquescript.psi.impl.TSPropertyElementImpl
-import org.lukasj.idea.torquescript.reference.TSFunctionReference
 import org.lukasj.idea.torquescript.util.TSTypeLookupService
 
 class TSMethodCallAnnotator : TSAnnotator() {
@@ -84,7 +79,7 @@ class TSMethodCallAnnotator : TSAnnotator() {
 
                 if (namespace != null) {
                     if (
-                        typeLookupService.getNamespaces(namespace, element.project)
+                        typeLookupService.getNamespaceInheritanceList(namespace, element.project)
                             .any {
                                 engineApiService.findMethod(it, functionName.text) != null
                             }